fix: round-trip spec-required scalar schema fields whose value is 0 (#706)
* fix: Helper properties in Schema classes don't get serialized by accident anymore.

* fix: round-trip spec-required scalar schema fields whose value is 0

sentinel-shim pattern. New `Constants.UnsetIndex` (= -1) and `Constants.UnsetByteLength` (= -1L) act as the "unset" marker. Each affected property keeps its non-nullable type but defaults to the sentinel; an internal `int?` shim returns null when the value is < 0 (omitted by the serializer) and 0+ when set. The negative half of the value space is spec-invalid for indices/counts/lengths, so we reuse it to encode absence without polluting the spec-valid range.

* fix: Corrected test reference values.

* chore: White space

* fix: Added missing TestAttribute

* fix: round-trip AnimationChannelTarget.Node value 0

* fix: round-trip BufferViewMeshoptExtension.Buffer value 0

* fix: Added bounds check to `GltfImport.GetAccessorSparseIndices`.

* fix: Added bounds check to `GltfImport.GetAccessorSparseValues`.


* fix: Null check


---------

+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Export/GltfWriter.cs
@@ -812,7 +812,7 @@ namespace GLTFast.Export
                 m_Gltf.Buffers = new List<Buffer> {
                     new Buffer {
                         Uri = string.IsNullOrEmpty(bufferPath) ? null : new UriValue(bufferPath),
-                        ByteLength = (uint) m_BufferStream.Length
+                        ByteLength = m_BufferStream.Length
                     }
                 };
             }
